+++ src/main/java/com/example/cab302tailproject/DAO/SqliteConnection.java
@@ -1,0 +1,24 @@
+package com.example.cab302tailproject.DAO;
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.SQLException;
+
+public class SqliteConnection {
+    private static Connection instance = null;
+
+    private SqliteConnection() {
+        String url = "jdbc:sqlite:Tail.db";
+        try {
+            instance = DriverManager.getConnection(url);
+        } catch (SQLException sqlEx) {
+            System.err.println(sqlEx);
+        }
+    }
+
+    public static Connection getInstance() {
+        if (instance == null) {
+            new SqliteConnection();
+        }
+        return instance;
+    }
+}