+++ src/main/java/com/example/cab302tailproject/TeacherGenerateApplication.java
@@ -2,7 +2,6 @@
 
 package com.example.cab302tailproject;
 
-import com.example.cab302tailproject.controller.LessonGenController;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
@@ -10,15 +9,15 @@ import javafx.stage.Stage;
 
 import java.io.IOException;
 
-public class LessonGenApplication extends Application {
+public class TeacherGenerateApplication extends Application {
     public static final String TITLE = "Worksheet and Lesson Plan Generator";
     public static final int WIDTH = 900;
     public static final int HEIGHT = 600;
 
     @Override
     public void start(Stage stage) throws IOException {
-        FXMLLoader fxmlLoader = new FXMLLoader(LessonGenApplication.class.getResource(
-                "lesson-gen-view.fxml"));
+        FXMLLoader fxmlLoader = new FXMLLoader(TeacherGenerateApplication.class.getResource(
+                "lesson_generator-teacher.fxml"));
         Scene scene = new Scene(fxmlLoader.load());
         stage.setTitle(TITLE);
         stage.setScene(scene);