+++ src/main/java/com/example/cab302tailproject/controller/LoginController.java
@@ -1,6 +1,7 @@
 package com.example.cab302tailproject.controller;
 
-import com.example.cab302tailproject.HelloApplication;
+import com.example.cab302tailproject.DAO.*;
+import com.example.cab302tailproject.TailApplication;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
 import javafx.stage.Stage;
@@ -13,6 +14,7 @@ import javafx.scene.control.TextField;
 import javafx.scene.control.PasswordField;
 import javafx.event.ActionEvent;
 
+
 public class LoginController {
 
     @FXML
@@ -20,33 +22,72 @@ public class LoginController {
     @FXML
     private Button registerPageButton;
     @FXML
-    private PasswordField loginPasswordTextField;
+    private PasswordField loginPasswordField;
     @FXML
     private TextField loginEmailTextField;
     @FXML
 
+    private TeacherDAO teacherDao;
+    private StudentDAO studentDAO;
+    public LoginController() {
+        teacherDao = new SqliteTeacherDAO();
+        studentDAO = new SqlStudentDAO();
+    }
+
     //Precondition -
     public boolean isEmpty(ActionEvent event) {
-        if (!loginEmailTextField.getText().isBlank() && !loginPasswordTextField.getText().isBlank()) {
+        if (!loginEmailTextField.getText().isBlank() && !loginPasswordField.getText().isBlank()) {
             return true;
         }
         return false;
     }
 
+    /**
+     * Button used to take a user to registration page to register for the application.
+     * @throws IOException
+     */
+
     @FXML
     protected void onRegistrationButtonClick() throws IOException {
         Stage stage = (Stage) registerPageButton.getScene().getWindow();
-        FXMLLoader fxmlLoader = new FXMLLoader(HelloApplication.class.getResource("RegistrationPage.fxml"));
-        Scene scene = new Scene(fxmlLoader.load(), HelloApplication.WIDTH, HelloApplication.HEIGHT);
+        FXMLLoader fxmlLoader = new FXMLLoader(TailApplication.class.getResource("registration_page.fxml"));
+        Scene scene = new Scene(fxmlLoader.load(), TailApplication.WIDTH, TailApplication.HEIGHT);
         stage.setScene(scene);
     }
+
+    /**
+     * Button which checks the validity of the login attempt against the data store in the Sqlite database and then
+     * the user to the appropriate view.
+     * @throws IOException
+     */
 
     @FXML
     protected void onLoginButtonClick() throws IOException {
-        Stage stage = (Stage) loginButton.getScene().getWindow();
-        FXMLLoader fxmlLoader = new FXMLLoader(HelloApplication.class.getResource("Main.fxml"));
-        Scene scene = new Scene(fxmlLoader.load(), HelloApplication.WIDTH, HelloApplication.HEIGHT);
-        stage.setScene(scene);
+        if (checkLogin()) {
+            Stage stage = (Stage) loginButton.getScene().getWindow();
+            FXMLLoader fxmlLoader = new FXMLLoader(TailApplication.class.getResource("lesson_generator-teacher.fxml"));
+            Scene scene = new Scene(fxmlLoader.load(), TailApplication.WIDTH, TailApplication.HEIGHT);
+            stage.setScene(scene);
+        }
     }
 
+    /**
+     * Boolean function that checks whether the user has registered by using the inputted email and password and checking
+     * it against the database.
+     * @return true if the credentials inputted are valid.
+     */
+    private boolean checkLogin() {
+        String email = loginEmailTextField.getText();
+        String password = loginPasswordField.getText();
+
+        return tryLogin(teacherDao, email, password)
+                || tryLogin(studentDAO, email, password);
+    }
+
+    private boolean tryLogin(UserDAO dao, String email, String password) {
+        return dao.checkEmail(email) && dao.checkPassword(email, password);
+    }
+
+
 }
+