+++ src/main/java/com/example/cab302tailproject/model/Student.java
@@ -1,0 +1,11 @@
+package com.example.cab302tailproject.model;
+
+public class Student extends User {
+    public Student(String firstName, String lastName, String email, String password) {
+        super(firstName, lastName, email, password);
+    }
+    public Student(String firstName, String lastName, String email) {
+        this(firstName, lastName, email, null);
+    }
+
+}