+++ src/main/java/com/example/cab302tailproject/model/User.java
@@ -1,0 +1,35 @@
+package com.example.cab302tailproject.model;
+
+public class User {
+    // Define Attributes for a user
+    protected String firstName;
+    protected String lastName;
+    protected String email;
+    protected String password;;
+
+    public User(String firstName, String lastName, String email, String password) {
+        this.firstName = firstName;
+        this.lastName = lastName;
+        this.email = email;
+        this.password = password;
+    }
+    // Getters
+    public String getFirstName() { return firstName; }
+    public String getLastName() { return lastName; }
+    public String getEmail() { return email; }
+    public String getPassword() { return password; }
+
+    // Setters
+    public void setFirstName(String firstName) { this.firstName = firstName; }
+    public void setLastName(String lastName) { this.lastName = lastName; }
+    public void setEmail(String email) { this.email = email; }
+    public void setPassword(String password) { this.password = password; }
+
+    public void login(){
+
+    }
+    public void logout(){
+
+    }
+}
+